--- PRUEBA/ordenacion.h
@@ -1,14 +1,0 @@
-#ifndef ORDENACION_H
-
-#define ORDENACION_H
-
-//Ordena usando  la funcion Shellsort
-void shellsort(int* Arr_crlV, int n );
-
-//Ordena usando la funcion Quicksort
-void QuickSort(int* Arr_crlV, int inicio, int final); 
-
-//Ordena usando la funcion de Insercion Binaria
-void InsercionBinaria(int* Arr_crlV, int n);
-
-#endif
