+++ Programa Busquedas - Jhonel G.A.P/Makefile.win
@@ -4,8 +4,8 @@
 CPP      = g++.exe
 CC       = gcc.exe
 WINDRES  = windres.exe
-OBJ      = main.o busqueda.o ordenacion.o utilidades.o
-LINKOBJ  = main.o busqueda.o ordenacion.o utilidades.o
+OBJ      = main.o busqueda.o utilidades.o
+LINKOBJ  = main.o busqueda.o utilidades.o
 LIBS     = -L"C:/Program Files/Dev-Cpp/MinGW64/lib" -L"C:/Program Files/Dev-Cpp/MinGW64/x86_64-w64-mingw32/lib" -static-libgcc
 INCS     = -I"C:/Program Files/Dev-Cpp/MinGW64/include" -I"C:/Program Files/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"C:/Program Files/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include"
 CXXINCS  = -I"C:/Program Files/Dev-Cpp/MinGW64/include" -I"C:/Program Files/Dev-Cpp/MinGW64/x86_64-w64-mingw32/include" -I"C:/Program Files/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include" -I"C:/Program Files/Dev-Cpp/MinGW64/lib/gcc/x86_64-w64-mingw32/4.9.2/include/c++"
@@ -29,9 +29,6 @@ main.o: main.cpp
 
 busqueda.o: busqueda.cpp
 	$(CPP) -c busqueda.cpp -o busqueda.o $(CXXFLAGS)
-
-ordenacion.o: ordenacion.cpp
-	$(CPP) -c ordenacion.cpp -o ordenacion.o $(CXXFLAGS)
 
 utilidades.o: utilidades.cpp
 	$(CPP) -c utilidades.cpp -o utilidades.o $(CXXFLAGS)
