+++ Programa Busquedas - Jhonel G.A.P/busqueda.h
