+++ Programa Busquedas - Jhonel G.A.P/proyectoFinalComparaciones.layout
@@ -1,10 +1,10 @@
 [Editors]
-Order=0,1,3,5,2,4,6
+Order=
 Focused=3
 [Editor_0]
 CursorCol=18
 CursorRow=205
-TopLine=163
+TopLine=162
 LeftChar=1
 [Editor_1]
 CursorCol=1
@@ -19,7 +19,7 @@ LeftChar=1
 [Editor_3]
 CursorCol=1
 CursorRow=89
-TopLine=43
+TopLine=4
 LeftChar=1
 [Editor_4]
 CursorCol=7
