+++ Programa Busquedas - Jhonel G.A.P/utilidades.cpp
@@ -7,20 +7,6 @@
 #include <iomanip> 
 using namespace std;
 
-//Funcion para generar arreglo aleatorio
-void generarArreglo(int* arr, int n){
-    srand(time(0));
-
-    for(int i=0; i<n; ++i)
-        arr[i] = rand() % 1000000 + 1;
-}
-//Funcion para copiar arreglo aleatorio
-void copiarArr(int* Arr_crlV, const int* Arr_crlC, int n){
-    for(int i=0;i<n;i++){
-        Arr_crlV[i]=Arr_crlC[i]; //El arreglo copiado sera igual original 
-    }
-}
-//Funcion para imprimir arreglo
 void imprimirArr(const int* arr, int n){
     for(int i=0; i<n ; i++){
         std::cout << arr[i]<<" ";
@@ -28,15 +14,10 @@ void imprimirArr(const int* arr, int n){
     std::cout<<std::endl;
 }
 
-void mostrarArreglo(int a[], int n) {
-    for (int i = 0; i < n; i++) {
-        cout << a[i] << " ";
-    }
-}
 
 void imprimirComparacion(Metodos tiempo[], Metodos promedio[], int n, int c, int datos, int busquedas, int elementoBuscar[], int r[]) {
 	cout << endl << endl;
-    cout << "C O M P A R A C I O N    " << c + 1 << endl << endl // Cambiado c a c + 1 para mostrar correctamente el índice
+    cout << "C O M P A R A C I O N    " << c + 1 << endl << endl
          << "Elemento a buscar: Aletorio por busqueda." << endl
          << "Arreglo: Sin repetir, aleatorio." << endl 
          << "Total de datos: " << datos << endl
@@ -44,16 +25,44 @@ void imprimirComparacion(Metodos tiempo[], Metodos promedio[], int n, int c, int
                 cout << "\tBUSQUEDA BINARIA" << "\t\t\t" << "BUSQUEDA EXPONENCIAL" << endl << endl;
     for (int i = 0; i < n; i++) {
 
-        cout << /*" (" << elementoBuscar[i] << ") " << */ "Busqueda " << i + 1 << " (" << elementoBuscar[i] << "):" << fixed << setprecision(7) << tiempo[i].busquedaBinaria << " ms" << "\t\t\t"  << tiempo[i].busquedaExponencial << " ms" << endl;
+        cout << "Busqueda " << i + 1 << " (" << elementoBuscar[i] << "):" 
+		<< fixed << setprecision(7) << tiempo[i].busquedaBinaria << " ms" << "\t\t\t"  
+		<< tiempo[i].busquedaExponencial << " ms" << endl;
     }
     cout << endl;
     cout << "\tPromedio: " << promedio[c].busquedaBinaria 
          << "\t\t\t" << "Promedio: " << promedio[c].busquedaExponencial << endl;
 }
 
+
 void mezclarArreglo(int* arreglo, int tamano) {
     default_random_engine engine(static_cast<unsigned int>(time(0)));
     shuffle(arreglo, arreglo + tamano, engine);
 }
 
+void arrSucesivo(int* a, int n){
+	for (int i = 0; i < n; i++) {
+        a[i] = i;
+    }
+}
+
+void shellsort(int* Arr_crlV, int n){
+    int k=n;
+    while(k>1){
+        k = k/2;
+        bool cen=true;
+        while(cen){
+            cen=false;
+            for(int i=0; i + k < n ; i++){
+                if(Arr_crlV[i] > Arr_crlV[i+k]){
+                    int aux=Arr_crlV[i];
+
+                    Arr_crlV[i]=Arr_crlV[i+k];
+                    Arr_crlV[i+k]=aux;
+                    cen=true;
+                }
+            }
+        }
+    }
+}
